Restructure and introduce a standalone config
